requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
pybars3==0.9.7
PyMeta3==0.5.1
requests==2.32.3
urllib3==2.2.3
